Moved server/pg.js to server/models/pg.js
